Added a function to print connection stats
- Added the `printStats()` function that prints some info like the local
and remote IPs and ports that were used to establish a P2P connection
- Also set up the script such that it tries to get the peerjsServerId
for the server from a config.json file. If it fails then it asks the
user to enter it manually. This will be useful if the user ever gets
annoyed of manually entering the peerjsServerId again and again